map/ply-crawler/requirements.txt
psycopg2-binary==2.9.9
pymeshlab==2023.12.post1
watchdog==4.0.1